# docs/openapi/lucy-openapi.yaml
openapi: 3.0.3
info:
  title: Backend API Documentation
  description: OpenAPI contract definitions mapped from controllers.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Server
paths:
...
  /api/mentor/rooms/mentor/{hostUserId}:
    get:
      tags:
        - Mentor Room
      summary: List rooms hosted by mentor
      parameters:
        - name: hostUserId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/mentor/rooms/{roomId}/end:
    post:
      tags:
        - Mentor Room
      summary: End a room session
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Room session ended
  /api/mentor/rooms/{roomId}/participants:
    get:
      tags:
        - Mentor Room
      summary: Get list of participants in room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/mentor/rooms/{roomId}/open:
    post:
      tags:
        - Mentor Room
      summary: Open room session
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Room opened
  /api/mentor/rooms/{roomId}/sublevels/{subLevelId}/complete:
    post:
      tags:
        - Mentor Room
      summary: Mark sub-level complete
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
        - name: subLevelId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sub-level completed

  # ==========================================
  # LEARNER ROOM CONTROLLER
  # ==========================================
  /api/learner/rooms/{roomId}/join:
    post:
      tags:
        - Learner Room
      summary: Join a room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Joined room
  /api/learner/rooms/{roomId}/leave:
    post:
      tags:
        - Learner Room
      summary: Leave a room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Left room
  /api/learner/rooms/{roomId}/mic:
    patch:
      tags:
        - Learner Room
      summary: Toggle microphone
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Microphone toggled
  /api/learner/rooms/{roomId}/hand-raise:
    patch:
      tags:
        - Learner Room
      summary: Toggle hand raise status
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Hand raise status toggled
  /api/learner/rooms/{roomId}/participants:
    get:
      tags:
        - Learner Room
      summary: List participants
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK

  # ==========================================
  # PINNED MATERIAL CONTROLLER
  # ==========================================
...